+++ src/micm/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(musica
   PRIVATE
     micm.cpp
-    parse.cpp
+    v0_parse.cpp
+    v1_parse.cpp
 )
